BSML使用におきかえ、OWN/HOST切り替え追加、BeatSaver開く追加

+++ MultiplayerLevelInformation/Utils/BeatmapUtils.cs
@@ -4,24 +4,24 @@ namespace MultiplayerLevelInformation.Utils
 {
     internal static class BeatmapUtils
     {
-        static float hjd(float bpm, float njs, float offset)
+        static double hjd(double bpm, double njs, double offset)
         {
             var num = 60f / bpm;
-            var hjd = 4f;
+            var hjd = 4d;
             while (njs * num * hjd > 17.999f)
                 hjd /= 2f;
 
             hjd += offset;
 
-            return Math.Max(hjd, 0.25f);
+            return Math.Max(hjd, 0.25d);
         }
 
-        public static float GetJd(float bpm, float njs, float offset)
+        public static double GetJd(double bpm, double njs, double offset)
         {
             return njs * (60f / bpm) * hjd(bpm, njs, offset) * 2;
         }
 
-        public static float GetRT(float bpm, float njs, float offset)
+        public static double GetRT(double bpm, double njs, double offset)
         {
             return GetJd(bpm, njs, offset) / (2 * njs);
         }
